+++ examples/codedb/dispatch.rb
@@ -1,0 +1,6 @@
+class CodedbJob
+  def self.run(n)
+    n * 2
+  end
+end
+$codedb_dispatch = CodedbJob.run(21)